% Generated by roxygen2 (4.0.2): do not edit by hand
\name{donorscape_report}
\alias{donorscape_report}
\title{Input a list of entity ids and output a csv file ready for use with the DonorScape tool}
\usage{
donorscape_report(ids, include_deceased = FALSE, household = TRUE,
  file = NULL)
}
\arguments{
\item{ids}{A list of entity ids}

\item{include_deceased}{Should deceased individuals be included?}

\item{household}{Should there be one line per household?}

\item{file}{Where should output go?}
}
\description{
Input a list of entity ids and output a csv file ready for use with the DonorScape tool
}
\details{
By default, deceased individuals are exclused.
By default, output is one line per household.
If no file location is indicated, the output will be returned within R.
}

